# docs/static/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Singularity API
  description: "Welcome to the *Singularity*! \n\nThis documentation covers the ready-to-use\
    \ endpoints \nfor authentication, user management, and content management that\
    \ come pre-configured with \n*Singularity*.\n\n**Features at a Glance**:\n\n-\
    \ **Authentication & User Management:** Includes JWT auth with refresh tokens,\
    \ 2FA, \n  email verification, and role-based access control.\n  \n- **Data &\
    \ Content Management:** Endpoints for interacting with a content management \n\
    \  system that supports multi-language content, configurable tagging, and built-in\
    \ \n  access control. A prebuilt `Article` class is also provided.\n  \n- **Performance\
    \ & Security:** The API is built with a focus on production-readiness, \n  including\
    \ rate limiting and secret manager integration with automated key rotation.\n\n\
    This API is a foundation for building robust backend services, microservices,\
    \ or \nfull-stack applications.\n\n**Note:** Some endpoints require explicit configuration\
    \ to be enabled."
  contact:
    name: Stereov
    url: https://stereov.io
    email: contact@stereov.io
  license:
    name: GPLv3 License
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
- url: https://singularity.stereov.io
  description: Demo server
- url: http://localhost:8000
  description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      description: "Authenticates a user with `email` and `password`.\n\n**Optional\
        \ session data:**\n- The `session` object can be included in the request body.\n\
        - Inside the `session` object, you can provide the following optional fields:\n\
        \    - `browser`: The name of the browser used (e.g., \"Chrome\", \"Firefox\"\
        ).\n    - `os`: The operating system of the device (e.g., \"Windows\", \"\
        macOS\", \"Android\").\n\nThis information helps users identify and manage\
        \ authorized sessions, improving overall account security.\n\n**2FA:**\n\n\
        If the user enabled 2FA, the user will not be authenticated immediately. \n\
        Instead, a [`TwoFactorAuthenticationToken`](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)\
        \ \nis set as HTTP-only cookie and returned in the response body\nif [header\
        \ authentication](https://singularity.stereov.io/docs/guides/auth/authentication#header-authentication)\
        \ is enabled.\n\nYou can complete the login through the endpoint [`POST /api/auth/2fa/login`](https://singularity.stereov.io/docs/api/complete-login).\n\
        \n**Locale:**\n\nA locale can be specified for this request. \nThis will be\
        \ used for the email 2FA code if this method is enabled for the user.\nYou\
        \ can learn more about 2FA through email [here](/docs/guides/auth/two-factor#email).\n\
        \nIf no locale is specified, the applications default locale will be used.\n\
        You can learn more about configuring the default locale [here](https://singularity.stereov.io/docs/guides/configuration).\n\
        \n**Tokens:**\n- If 2FA is disabled and the request is successful, \n  [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)\
        \ \n  and [`RefreshToken`](https://singularity.stereov.io/docs/guides/auth/tokens#refresh-token)\
        \ \n  will automatically be set as HTTP-only cookies.\n  If [header authentication](https://singularity.stereov.io/docs/guides/auth/authentication#header-authentication)\
        \ is enabled,\n  [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)\
        \ and \n  [`RefreshToken`](https://singularity.stereov.io/docs/guides/auth/tokens#refresh-token)\
        \ \n  will be returned in the response body and can be used as \n  bearer\
        \ tokens in the authorization header for upcoming requests."
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/authentication#login
      operationId: login
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Authentication successful. Returns tokens and user details.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "304":
          description: User is already authenticated. Authenticated session state
            has not changed since last request.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Trying to log in user that did not set up authentication using
            password.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
      - accessTokenCookie: []
...
  /api/auth/step-up:
    post:
      tags:
      - Authentication
      summary: Step-Up
      description: "Requests [step-up authentification](https://singularity.stereov.io/docs/guides/auth/authentication#step-up).\
        \ \nThis re-authentication is required by critical endpoints.\n\n**2FA:**\n\
        \nIf the user enabled 2FA, the step-up will not be granted immediately. \n\
        Instead, a [`TwoFactorAuthenticationToken`](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)\
        \ \nis set as HTTP-only cookie and returned in the response body\nif [header\
        \ authentication](https://singularity.stereov.io/docs/guides/auth/authentication#header-authentication)\
        \ is enabled.\n\nYou can complete the step-up through the endpoint [`POST\
        \ /api/auth/2fa/step-up`](https://singularity.stereov.io/docs/api/complete-step-up).\n\
        \n**Request Body:**\n\nWhen requesting a step-up for a [`GUEST`](https://singularity.stereov.io/docs/guides/auth/roles#guests)\n\
        there is no way to authenticate the user.\nTherefore, no request body is required\
        \ in this case.\n\nIf you request a step-up for a regular [`USER`](https://singularity.stereov.io/docs/guides/auth/roles#users),\n\
        it will result in a `400 - BAD REQUEST`.\n\n**Locale:**\n\nA locale can be\
        \ specified for this request. \nThis will be used for the email 2FA code if\
        \ this method is enabled for the user.\nYou can learn more about 2FA through\
        \ email [here](/docs/guides/auth/two-factor#email).\n\nIf no locale is specified,\
        \ the applications default locale will be used.\nYou can learn more about\
        \ configuring the default locale [here](https://singularity.stereov.io/docs/guides/configuration).\n\
        \n**Tokens:**\n- Requires a valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token).\n\
        - If 2FA is disabled and the request is successful, [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)\n\
        \  will automatically be set as HTTP-only cookie.\n  If [header authentication](https://singularity.stereov.io/docs/guides/auth/authentication#header-authentication)\
        \ is enabled,\n  the [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)\n\
        \  will be returned in the response body and can be used to authorized critical\
        \ requests."
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/authentication#step-up
      operationId: stepUp
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpRequest"
      responses:
        "200":
          description: Logout successful.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StepUpResponse"
        "400":
          description: Trying to request step-up for user that authenticated only
            via OAuth2 providers or missing request body for authenticated users.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
      - accessTokenCookie: []
...
  /api/auth/2fa/login:
    post:
      tags:
      - Two-Factor Authentication
      summary: Complete Login
      description: "Complete second factor for login.\n\nYou can learn more about\
        \ the login flow [here](https://singularity.stereov.io/docs/guides/auth/authentication#login).\n\
        \n**Requirements:**\n- User authenticated using their password through [`POST\
        \ /api/auth/login`](https://singularity.stereov.io/docs/api/login).\n- At\
        \ least one of [email](https://singularity.stereov.io/docs/guides/auth/two-factor#email)\
        \ or [TOTP](https://singularity.stereov.io/docs/guides/auth/two-factor#totp)\
        \ as 2FA methods should be enabled.\n- An `email` or `totp` 2FA code is present\
        \ for an enabled 2FA method.\n  Check out [email](https://singularity.stereov.io/docs/guides/auth/two-factor#email)\
        \ and [TOTP](https://singularity.stereov.io/docs/guides/auth/two-factor#totp)\n\
        \  to learn how to retrieve a 2FA code.\n  \n**Optional session data:**\n\
        - The `session` object can be included in the request body.\n- Inside the\
        \ `session` object, you can provide the following optional fields:\n    -\
        \ `browser`: The name of the browser used (e.g., \"Chrome\", \"Firefox\").\n\
        \    - `os`: The operating system of the device (e.g., \"Windows\", \"macOS\"\
        , \"Android\").\n\nThis information helps users identify and manage authorized\
        \ sessions, improving overall account security.\n\n**Tokens:**\n- A valid\
        \ [`TwoFactorAuthenticationToken`](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)\n\
        \  is required. This token will be set automatically as HTTP-only cookie through\
        \ [`POST /api/auth/login`](https://singularity.stereov.io/docs/api/login)\n\
        \  or can be retrieved from the response and set as header manually if [header\
        \ authentication](https://singularity.stereov.io/docs/guides/authentication#header-authentication)\
        \ \n  is enabled.\n- If this action is successful, [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)\
        \ and \n  [`RefreshToken`](https://singularity.stereov.io/docs/guides/auth/tokens#refresh-token)\
        \ \n  will automatically be set as HTTP-only cookies.\n  If [header authentication](https://singularity.stereov.io/docs/guides/auth/authentication#header-authentication)\
        \ is enabled,\n  [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)\
        \ and \n  [`RefreshToken`](https://singularity.stereov.io/docs/guides/auth/tokens#refresh-token)\n\
        \  will be returned in the response body and can be used as \n  bearer tokens\
        \ in the authorization header for upcoming requests."
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/two-factor
      operationId: completeLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteLoginRequest"
        required: true
      responses:
        "200":
          description: Information about the user and the tokens if header authentication
            is enabled.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: No 2FA code for an enabled 2FA method was provided or 2FA is
            disabled.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired `TwoFactorAuthenticationToken`.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - twoFactorAuthenticationTokenHeader: []
      - twoFactorAuthenticationTokenCookie: []
  /api/auth/2fa/step-up:
    post:
      tags:
      - Two-Factor Authentication
      summary: Complete Step-Up
      description: |-
        Perform second factor for step-up.

        You can learn more about the step-up flow [here](https://singularity.stereov.io/docs/guides/auth/authentication#step-up).

        **Requirements:**
        - User authenticated using their password through [`POST /api/auth/step-up`](https://singularity.stereov.io/docs/api/step-up).
        - At least one of [email](https://singularity.stereov.io/docs/guides/auth/two-factor#email) or [TOTP](https://singularity.stereov.io/docs/guides/auth/two-factor#totp) as 2FA methods should be enabled.
        - An `email` or `totp` 2FA code is present for an enabled 2FA method.
          Check out [email](https://singularity.stereov.io/docs/guides/auth/two-factor#email) and [TOTP](https://singularity.stereov.io/docs/guides/auth/two-factor#totp)
          to learn how to retrieve a 2FA code.

        **Tokens:**
        - A valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token) is required.
        - A valid [`TwoFactorAuthenticationToken`](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)
          is required. This token should match user and session contained in the `AccessToken`.
        - If 2FA is disabled and the request is successful, [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)
          will automatically be set as HTTP-only cookie.
          If [header authentication](https://singularity.stereov.io/docs/guides/auth/authentication#header-authentication) is enabled,
          the [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)
          will be returned in the response body and can be used to authorized critical requests.
      externalDocs:
        url: https://singularity.stereov.io/docs/auth/two-factor
      operationId: completeStepUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteStepUpRequest"
        required: true
      responses:
        "200":
          description: The token if header authentication is enabled.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StepUpResponse"
        "400":
          description: No 2FA code for an enabled 2FA method was provided or 2FA is
            disabled.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired `TwoFactorAuthenticationToken` or invalid
            or expired 2FA code.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
        twoFactorAuthenticationTokenHeader: []
      - accessTokenCookie: []
        twoFactorAuthenticationTokenCookie: []
  /api/auth/2fa/preferred-method:
    post:
      tags:
      - Two-Factor Authentication
      summary: Change Preferred 2FA Method
      description: "Change the preferred 2FA method.\n\nYou can learn more about 2FA\
        \ methods [here](https://singularity.stereov.io/docs/guides/auth/two-factor).\n\
        \n**Requirements:**\n- The user can authenticate using password. 2FA will\
        \ not work with OAuth2. \n  The OAuth2 provider will validate the second factor\
        \ if the user enabled it for the provider.\n- At least one of [email](https://singularity.stereov.io/docs/guides/auth/two-factor#email)\
        \ or [TOTP](https://singularity.stereov.io/docs/guides/auth/two-factor#totp)\
        \ as 2FA methods should be enabled.\n\n**Tokens:**\n- A valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)\
        \ is required.\n- A valid [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)\n\
        \  is required. This token should match user and session contained in the\
        \ `AccessToken`."
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/two-factor#changing-the-preferred-method
      operationId: changePreferredTwoFactorMethod
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferredTwoFactorMethodRequest"
        required: true
      responses:
        "200":
          description: Updated user information.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: This 2FA method is disabled.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: '`AccessToken` or `StepUpToken` are invalid or expired or invalid
            or expired 2FA code.'
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
        stepUpTokenHeader: []
      - accessTokenCookie: []
        stepUpTokenCookie: []
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required:
      - error
      - status
      - timestamp
...
    StepUpRequest:
      type: object
      properties:
        password:
          type: string
      required:
      - password
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        session:
          $ref: "#/components/schemas/SessionInfoRequest"
      required:
      - email
      - password
...
  securitySchemes:
    twoFactorAuthenticationTokenCookie:
      type: apiKey
      description: "Token for successful authentication with email and password, indicating\
        \ 2FA is required.You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)."
      name: two_factor_authentication_token
      in: cookie
    accessTokenHeader:
      type: http
      description: "Access token for user authentication. You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)."
      scheme: bearer
      bearerFormat: JWT
    accessTokenCookie:
      type: apiKey
      description: "Access token for user authentication.You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)."
      name: access_token
      in: cookie
    stepUpTokenCookie:
      type: apiKey
      description: "Token for step up authentication allowing access of secure resources.You\
        \ can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)."
      name: step_up_token
      in: cookie
    twoFactorAuthenticationTokenHeader:
      type: apiKey
      description: "Token for successful authentication with email and password, indicating\
        \ 2FA is required.You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)."
      name: X-Two-Factor-Authentication-Token
      in: header
    refreshTokenHeader:
      type: http
      description: "Refresh token to request new access tokens.You can learn more\
        \ [here](https://singularity.stereov.io/docs/guides/auth/tokens#refresh-token)."
      scheme: bearer
      bearerFormat: JWT
    stepUpTokenHeader:
      type: apiKey
      description: "Token for step up authentication allowing access of secure resources.You\
        \ can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)."
      name: X-Step-Up-Token
      in: header
...